refactor(content): switch from a svelte timer component to store
